% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_stats.r
\name{bal_loci}
\alias{bal_loci}
\title{Balance per-locus sample size between two populations.}
\usage{
bal_loci(data)
}
\value{
A modified version of the input dataframe in hierfstat format.
}
\description{
Takes a hierfstat input file and equalizes sample size independently for each
locus by randomly removing genotypes from the sample with higher n. Returns
the modified hierfstat input file. If more than two population samples are
present in the input file, then only the first two listed pop samples are
processed.
}
\examples{
test <- bal_loci(dat.sim)
}
