% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.sim}
\alias{dat.sim}
\title{Simulated genotype data.}
\format{A dataframe with 20 rows and 101 columns.}
\usage{
dat.sim
}
\description{
A dataset containing genotype data for two population samples, each
population consists of 10 individuals (rows). The population name for each
individual is in column 1 (="Pop"). Genotype data for each locus is in
columns (one locus per column). 100 loci are included and a random amount of
missing data (NA) has been incorporated for each locus. The data is in
hierfstat format and was generated using the function
hierfstat::sim.genot(size=10,nbal=2,nbloc=100,nbpop=2,N=c(1000,500),mig=0.001,mut=0.001,f=0).
}
\keyword{datasets}
