% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_stats.r
\name{Hs_calc}
\alias{Hs_calc}
\title{Gene diversity (Hs).}
\usage{
Hs_calc(data)
}
\value{
A matrix of calculated Hs values, with one row per locus and one
  column per population sample.
}
\description{
Calculates Nei's (1987) gene diversity (= Hs = expected heterozygosity
corrected for small sample size) for each locus and each population sample.
Code is modified from function hierfstat::basic.stats and relies on several
functions from the hierfstat library: pop.freq(), ind.count(), getal.b().
}
\examples{
test <- Hs_calc(dat.sim)
}
