% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fun.r
\name{plot_bal_loci}
\alias{plot_bal_loci}
\title{Plot difference in sample size for each locus.}
\usage{
plot_bal_loci(data, ...)
}
\value{
A lineplot showing difference in sample size for each locus.
}
\description{
Provides a quick visual check to compare missing data between two populations
and to verify that function bal_loci() has equalised sample sizes correctly.
Expects data in hierfstat format. #' If more than two population samples are
present in the input file, then only the first two listed pop samples are
processed. Plot can be modified using any additional arguments to plot() e.g.
add a title with the argument: main = "title text".
}
\examples{
test <- plot_bal_loci(dat.sim, main = "With missing data")
}
