% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy_fst.r
\name{rarefy_fst}
\alias{rarefy_fst}
\title{Population-specific FST with sample rarefaction.}
\usage{
rarefy_fst(data, rar.n = NULL, reps = 100)
}
\arguments{
\item{rar.n}{Rarefied sample size for each population. Must be equal or less
than the minimum population sample size. If not specified, rar.n defaults
to a value equal to the smallest sample size in the dataset.}

\item{reps}{Number of iterations to repeat the rarefied FST calculation. If
not specified, defaults to 100.}
}
\value{
Dataframe of FST estimates, where number of rows equal to iterations,
  and columns represent populations.
}
\description{
Randomly subsample each population in the dataset to a value (rar.n) set by
user, then calculate population-specific FST for each population sample using
hierfstat::betas. Expects data in hierfstat format.
}
\examples{
test <- rarefy_fst(dat.sim, rar.n = 5, reps = 100)
}
