% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{RunEnrichGO}
\alias{RunEnrichGO}
\title{Calculate gene enrichment by ORA method}
\usage{
RunEnrichGO(
  genes = NULL,
  org_db = "org.Hs.eg.db",
  ont = "BP",
  min_n = 3,
  max_n = 500,
  outdir = "."
)
}
\arguments{
\item{genes}{gene list}

\item{org_db}{org db}
}
\value{
enrich object
}
\description{
Calculate gene enrichment by ORA method
}
