% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{RunGSEA}
\alias{RunGSEA}
\title{Calculate gene enrichment by FGSEA method}
\usage{
RunGSEA(data = NULL, fgmt = "db.gmt", min_n = 3, max_n = 500)
}
\arguments{
\item{data}{data-frame}

\item{fgmt}{gmt file}
}
\value{
data frame
}
\description{
Calculate gene enrichment by FGSEA method
}
