% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{RunORA}
\alias{RunORA}
\title{Calculate gene enrichment by ORA method}
\usage{
RunORA(genes = NULL, fgmt = "db.gmt", min_n = 3, max_n = 500)
}
\arguments{
\item{genes}{gene list}

\item{fgmt}{gmt file}
}
\value{
enrich object
}
\description{
Calculate gene enrichment by ORA method
}
