% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_stats.r
\name{shuff_data}
\alias{shuff_data}
\title{Calculate null distribution of difference in gene diversity (Hs) between two
populations.}
\usage{
shuff_data(data, reps = 100, bal.loci = TRUE)
}
\arguments{
\item{reps}{Number of iterations to repeat the permutation. If not specified,
defaults to 100.}

\item{bal.loci}{Option to incorporate balancing of per-locus sample size into
the permutation (see function bal_loci). When bal.loci = TRUE, each
replicate of the permutation first balances sample sizes before gene
diversity calculation is performed. When bal.loci = FALSE, sample balancing
is skipped.}
}
\value{
A vector of values of length "reps". Each value represents the
  difference in gene diversity between two populations after individuals have
  been randomly permuted between the populations. This difference is
  calculated as Hs(pop2) - Hs(pop1).
}
\description{
Randomly permutes individuals between sample groups and calculates difference
in gene diversity (Hs) between the two randomised groups. Used to generate
null distribution of difference in gene diversity between two samples for
permutation testing
}
\examples{
test <- shuff_data(data = dat.sim, reps = 100, bal.loci = FALSE)
}
