% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{RunPathwayGroup}
\alias{RunPathwayGroup}
\title{Calculate gene enrichment by ORA method and pathway}
\usage{
RunPathwayGroup(
  df = NULL,
  fun = "enrichGO",
  species = "Homo sapiens",
  category = "C2",
  org_db = NULL,
  ont = "ALL",
  min_n = 3,
  max_n = 500,
  colors = c(`0` = "#FFA500", `1` = "#4B0082", `2` = "#20B2AA"),
  prefix = "pathway_group",
  size_category = 1,
  w = 8,
  h = 5,
  outdir = "."
)
}
\arguments{
\item{df}{dataframe}

\item{org_db}{org db}
}
\description{
Calculate gene enrichment by ORA method and pathway
}
